first part of changes from forks/iter/fort.13-14-15_safeties to forks/iter/MPI_barriers

+++ src/plotting/plane.f
@@ -71,7 +71,7 @@ C
           ENDIF
    10   CONTINUE
       ELSEIF (RL.GT.-10.) THEN
-        ILN=-RL
+        ILN=-INT(RL)
         DO 20 I=1,ILN
           IP=I+1
           DO 21 J=IP,ILN
